Refactor database service structure: rename and reorganize repositories, enhance error logging, and update API token handling
