+++ src/config/envs.ts
@@ -49,7 +49,7 @@ export const envs = {
   PORCENTAJE_IVA: envVars.PORCENTAJE_IVA,
 
   PAGOS_API_URL: envVars.PAGOS_API_URL,
-  PAGOS_API_TOKEN_TS: envVars.PAGOS_API_TOKEN_SS,
+  PAGOS_API_TOKEN_TS: envVars.PAGOS_API_TOKEN_TS,
   PAGOS_API_TOKEN_SS: envVars.PAGOS_API_TOKEN_SS,
   PAGOS_API_TOKEN_CS: envVars.PAGOS_API_TOKEN_CS,
 };
