+++ src/database/repositories/index.ts
@@ -1,0 +1,5 @@
+import { CuentasPagoApiService } from './cuentasPagoApi.repository';
+
+export const DatabaseRepositories = [CuentasPagoApiService];
+
+export * from './cuentasPagoApi.repository';
