--- src/database/services/index.ts
@@ -1,5 +1,0 @@
-import { CuentasPagoApiService } from './cuentasPagoApi.service';
-
-export const DatabaseTableServices = [CuentasPagoApiService];
-
-export * from './cuentasPagoApi.service';
